requirements.txt
cycler==0.10.0
joblib==1.0.1
keras~=2.4.3
kiwisolver==1.3.2
matplotlib~=3.4.2
numpy~=1.18.5
opencv-python~=4.5.1.48
pandas~=1.2.3
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
scikit-learn~=0.24.1
scipy==1.7.1
seaborn~=0.11.1
six==1.16.0
tensorboard==2.4.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
threadpoolctl==2.2.0

tensorflow~=2.2.2
pydicom~=2.2.2
tqdm~=4.62.3